fix: avoid use of WebGL2RenderingContext in webworker
This fixes Neuroglancer for firefox.
